The wrong file has been modified, I undo last change and apply them to the correct file
This commit correct the last commit modification because the wrong file has been edited, so the new modification has been applied to common.js file instead of share.js. The modification is about a missing scrollbar in firefox. An argument was missing in the window.open method. (see previous commit)